requirements.txt


#pip install python-dotenv==1.0.0 Flask==3.0.0 Flask-SQLAlchemy==3.1.1 Flask-Migrate==4.0.5 Flask-WTF==1.2.1 Flask-Login==0.6.3 WTForms==3.1.2 "email_validator>=2.0.0" Flask-Bootstrap==3.3.7.1 bleach==6.1.0 cryptography==42.0.5 pymysql==1.1.0 mysql-connector-python==8.0.33 mysqlclient==2.2.1 --no-cache-dir
#pip install requests
#pip install flask-login
# This file is used to install the required packages for the Flask application.
# It can be used with pip to install all dependencies at once.
# To install the packages, run:
# pip install -r requirements.txt
# or
# pip install -r requirements.txt --no-cache-dir
# This file is generated by pip freeze and should not be edited manually.
# If you need to add a new package, install it using pip and then run:
# pip freeze > requirements.txt
# to update this file with the new package and its version.
# Note: The versions specified here are for compatibility with the Flask application.
# Ensure that you have the correct versions of the packages to avoid compatibility issues.
# If you encounter any issues, check the package documentation for troubleshooting.
# This file is used to manage the dependencies of the Flask application.
# It is recommended to use a virtual environment to isolate the dependencies.
# You can create a virtual environment using:
# python -m venv venv
# and activate it using:
# source venv/bin/activate (Linux/Mac)
# venv\Scripts\activate (Windows)
# This will ensure that the packages are installed in the virtual environment and do not interfere with the system Python installation.



# requirements.txt
python-dotenv==1.0.0
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Login==0.6.3
WTForms==3.1.2
email_validator>=2.0.0
Flask-Bootstrap==3.3.7.1
bleach==6.1.0
cryptography==42.0.5
pymysql==1.1.0
mysql-connector-python==8.0.33
mysqlclient==2.2.1

